+++ R/internal.R
@@ -12,21 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-#' Retrieve Object Address
-#'
-#' Retrieve the pointer address of \code{x}.
-#'
-#' @param x, any R object, the object whose pointer address is retrieved.
-#'
-#' @return character scalar, the address of the object.
-#'
-#' @author Ron Triepels
-#' @keywords internal
-address <- function(x)
-{
-  .Call("address", x, PACKAGE = "cgraph")
-}
-
 #' Block Summation
 #'
 #' Divide a vector or array in consecutive blocks of \code{n} elements and sum the elements at each position in these blocks.
@@ -43,4 +28,51 @@ address <- function(x)
 bsum <- function(x, n = 1)
 {
   .Call("bsum", x, n, PACKAGE = "cgraph")
+}
+
+#' Approximate Gradients
+#'
+#' Differentiate node \code{x} with respect to node \code{y} by numerical differentiation.
+#'
+#' @param x character scalar, name of the node.
+#' @param y character scalar, name of the node.
+#' @param index numeric scalar, index of the target node that is differentiated. Defaults to the first element.
+#' @param eps numeric scalar, step size. Defaults to 1e-4.
+#'
+#' @note All nodes required to compute node \code{name} must have a value, or their value must be able to be computed at run-time.
+#'
+#' The graph is differentiation by the symmetric difference quotient. This method can only be used to differentiate scalars. In case the value of target node \code{name} is a vector or an array, argument \code{index} can be used to specify which element of the vector or array is differentiated. The caluclated gradient has the same shape as node \code{y}.
+#'
+#' Numerical differentiation is subject to estimation error and can be very slow. Therefore, this function should only be used for testing purposes.
+#'
+#' @return numeric vector or array, the derivative of \code{x} with respect to \code{y}.
+#'
+#' @author Ron Triepels
+#' @keywords internal
+approx_grad <- function(x, y, index = 1, eps = 1e-4)
+{
+  grad <- y.value <- val(y)
+
+  for(i in 1:length(grad))
+  {
+    y.value[i] <- y.value[i] + eps
+
+    set(y, y.value)
+
+    x.value1 <- val(x)[index]
+
+    y.value[i] <- y.value[i] - 2 * eps
+
+    set(y, y.value)
+
+    x.value2 <- val(x)[index]
+
+    y.value[i] <- y.value[i] + eps
+
+    set(y, y.value)
+
+    grad[i] <- (x.value1 - x.value2) / (2 * eps)
+  }
+
+  grad
 }